test(agent): isolate native glossary tsquery helper
Move punctuation sanitization into a dependency-free module so unit tests
do not require DATABASE_URL.
